% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_PWD.r
\name{multi_PWD}
\alias{multi_PWD}
\title{Multiple Instrument Weighted Deming Regression}
\usage{
multi_PWD(X, names, refine=1)
}
\arguments{
\item{X}{the \emph{n} by \emph{I} matrix of values reported by the \emph{I} instruments on the \emph{n} samples.}

\item{names}{\emph{optional} (default of NA) - labels of the \emph{I} instruments.}

\item{refine}{\emph{optional} (default of 1) - the number of resubstitutions used to refine lambda estimates.}
}
\value{
A list containing the following components:

\item{alpha}{the vector of \emph{I} estimated intercepts.}
\item{beta}{the vector of \emph{I} estimated slopes.}
\item{fity}{the \emph{n} by \emph{I} matrix of fitted values.}
\item{mu}{the vector of \emph{n} estimated true concentrations.}
\item{resi}{the \emph{n} by \emph{I} matrix of residuals.}
\item{rho}{the estimate of \eqn{\rho}.}
\item{sigma}{the estimate of \eqn{\sigma}.}
\item{kappa}{the estimate of \eqn{\kappa}.}
\item{lambda}{the vector of \emph{I} estimated \eqn{\lambda_i}.}
\item{L}{the -2 log likelihood}
\item{scalr}{the \emph{n} by \emph{I} matrix of scaled residuals.}
\item{whichmissing}{a logical vector identifying locations of row of X with missing values}
\item{results}{a printer-ready data frame of the fitted regression}
\item{names}{labels of the \emph{I} instruments.}
}
\description{
This code estimates the Rocke-Lorenzato precision profile parameters and the
regression of each instrument’s values on the underlying latent true
concentration.
}
\details{
The function optimizes the likelihood over the parameter \code{rho},
the ratio of the Rocke-Lorenzato parameters sigma to kappa: \eqn{\rho = \frac{\sigma}{\kappa}}.
Calls \code{multi_PWD_inner} to provide the estimates of the other parameters.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
n  <- 100
I  <- 4
alpha  <- c(0  , 0  , 2  , -2 )
beta   <- c(0.9, 1.1, 1.2, 0.8)
names   <- c("Inst_1", "Inst_2", "Inst_3", "Inst_4")
lambda <- c(1.8, 0.45, 0.9, 0.9)
rutlam <- sqrt(lambda)
sigma  <- 2
kappa  <- 0.08
true   <- 8*10^((0:(n-1))/(n-1))

# simulated data
set.seed(1039)
X     <- NULL
for (i in 1:I)  {
  truey <- alpha[i] + beta[i]*true
  X <- cbind(X, truey+rutlam[i]*(sigma*rnorm(n) + truey*kappa*rnorm(n)))
}

# fit with RL precision profile to estimate parameters
fit <- multi_PWD(X, names=names)

# results
summary(fit)

}
\references{
Hawkins DM and Kraker JJ. Multiple Instrument Methods Comparison
by Precision weighted Deming Regression,
on \emph{Arxiv} (2026) \url{doi:10.48550/arXiv.2607.11776}

.
}
\author{
Douglas M. Hawkins, Jessica J. Kraker \href{mailto:krakerjj@uwec.edu}{krakerjj@uwec.edu}
}
