% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_PWD_inner.r
\name{multi_PWD_inner}
\alias{multi_PWD_inner}
\title{Multiple Instrument Parameter Estimation from rho}
\usage{
multi_PWD_inner(X, rho, refine=1, eps=1e-9)
}
\arguments{
\item{X}{the \emph{n} by \emph{I} matrix of values reported by the \emph{I} instruments on the \emph{n} samples.}

\item{rho}{the input value of \eqn{\rho = \frac{\sigma}{\kappa}}.}

\item{refine}{\emph{optional} (default of 1) - the number of re-substitutions used to refine lambda estimates.}

\item{eps}{\emph{optional} (default of 1e-9) - a convergence criterion.}
}
\value{
A list containing the following components:

\item{alpha}{the vector of \emph{I} estimated intercepts \eqn{\alpha_i}.}
\item{beta}{the vector of \emph{I} estimated slopes \eqn{\beta_i}.}
\item{fity}{the \emph{n} by \emph{I} matrix of fitted values.}
\item{mu}{the vector of \emph{n} estimated true concentrations.}
\item{resi}{the \emph{n} by \emph{I} matrix of residuals.}
\item{rho}{the input value of \eqn{\rho}.}
\item{sigma}{the estimate of \eqn{\sigma}.}
\item{kappa}{the estimate of \eqn{\kappa}.}
\item{lambda}{the vector of \emph{I} estimated \eqn{\lambda_i}.}
\item{L}{the -2 log likelihood}
\item{scalr}{the \emph{n} by \emph{I} matrix of scaled residuals.}
}
\description{
This is primarily a workhorse routine that \code{multi_PWD} calls and, given a
value of \code{rho}, estimates the other model parameters.
}
\details{
Using the specified \eqn{\rho}, the routine calculates the estimates of
the regression parameters and \eqn{\lambda}.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
n  <- 100
I  <- 4
alpha  <- c(0  , 0  , 2  , -2 )
beta   <- c(0.9, 1.1, 1.2, 0.8)
names   <- c("Inst_1", "Inst_2", "Inst_3", "Inst_4")
lambda <- c(1.8, 0.45, 0.9, 0.9)
rutlam <- sqrt(lambda)
sigma  <- 2
kappa  <- 0.08
true   <- 8*10^((0:(n-1))/(n-1))

# simulated data
set.seed(1039)
X     <- NULL
for (i in 1:I)  {
  truey <- alpha[i] + beta[i]*true
  X <- cbind(X, truey+rutlam[i]*(sigma*rnorm(n) + truey*kappa*rnorm(n)))
}

}
\references{
Hawkins DM and Kraker JJ. Multiple Instrument Methods Comparison
by Precision weighted Deming Regression,
on \emph{Arxiv} (2026) \url{doi:10.48550/arXiv.2607.11776}

.
}
\author{
Douglas M. Hawkins, Jessica J. Kraker \href{mailto:krakerjj@uwec.edu}{krakerjj@uwec.edu}
}
