% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppwdeming-package.R
\docType{package}
\name{ppwdeming-package}
\alias{ppwdeming}
\alias{ppwdeming-package}
\title{ppwdeming: Precision Profile Weighted Deming Regression}
\description{
Weighted Deming regression, also known as "errors-in-variable" regression, is applied with suitable weights. Weights are modeled via a precision profile; functions are provided for implementing it in both known and unknown precision profile situations. The package provides tools for precision profile weighted Deming (PWD) regression. It covers two settings – one where the precision profiles are known either from external studies or from adequate replication of the X and Y readings, and one in which there is a plausible functional form for the precision profiles but the exact function must be estimated from the (generally singlicate) readings. The function set includes tools for: estimated standard errors (via jackknifing); standardized-residual analysis function with regression diagnostic tools for normality, linearity and constant variance; and an outlier analysis identifying significant outliers for closer investigation. Further information on mathematical derivations and applications is documented by DM Hawkins and JJ Kraker on Arxiv (2025, \href{https://arxiv.org/abs/2508.02888}{arXiv:2508.02888}).
}
\author{
\strong{Maintainer}: Jessica J. Kraker \email{krakerjj@uwec.edu}

Authors:
\itemize{
  \item Douglas M. Hawkins
}

}
\keyword{internal}
