% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWD_get_gh.r
\name{PWD_get_gh}
\alias{PWD_get_gh}
\title{Estimate of Variance Profile Functions (proportional)}
\usage{
PWD_get_gh(X, Y, lambda = 1,
           rho=lifecycle::deprecated(),
           alpha=lifecycle::deprecated(), beta=lifecycle::deprecated(),
           mu=lifecycle::deprecated(),
           quad=FALSE, epsilon = 1e-8,
           printem=lifecycle::deprecated())
}
\arguments{
\item{X}{the vector of predicate readings.}

\item{Y}{the vector of test readings.}

\item{lambda}{\emph{optional} (default of 1) - the ratio of the \code{X} to
the \code{Y} precision profile.}

\item{rho}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{rho = numvalue} initialization is no longer implemented in algorithm.}

\item{alpha}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{alpha = numvalue} initialization is no longer implemented in algorithm.}

\item{beta}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{beta = numvalue} initialization is no longer implemented in algorithm.}

\item{mu}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{mu = numvector} initialization is no longer implemented in algorithm.}

\item{quad}{\emph{optional} (default of FALSE) - logical, selects fitting a linear or a quadratic regression.}

\item{epsilon}{\emph{optional} (default of 1.e-8) - convergence tolerance limit.}

\item{printem}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{printem = TRUE} is no
longer supported; separate summary functions are provided for printing output.}
}
\value{
A list containing the following components:

\item{alpha }{the fitted intercept}
\item{beta }{the fitted slope}
\item{gamma}{the fitted coefficient of the square: if quad is FALSE, the value is zero}
\item{quad}{logical, FALSE for linear or TRUE for quadratic regression}
\item{fity }{the vector of predicted Y}
\item{mu }{the vector of estimated latent true values}
\item{resi }{the vector of residuals}
\item{sigma }{the estimate of the Rocke-Lorenzato \eqn{\sigma}}
\item{kappa }{the estimate of the Rocke-Lorenzato \eqn{\kappa}}
\item{L }{the -2 log likelihood L}
}
\description{
This code estimates the variance profiles, assumed proportional, of
the Rocke-Lorenzato form;
also provides the resulting weighted Deming fit and residuals.
}
\details{
This workhorse routine optimizes the likelihood in the \strong{unknown} \emph{g}, \emph{h}
setting over its \emph{n}+4 or \emph{n}+5 parameters:
the two Rocke-Lorenzato precision profile parameters \eqn{\sigma}
and \eqn{\kappa}, the intercept \eqn{\alpha} and slope \eqn{\beta}
(and coefficien of the squared term \eqn{\gamma}, if quadratic),
and the \emph{n} latent true concentrations \eqn{\mu_i}.

That is, the assumed forms are:
\itemize{
\item predicate precision profile model: \eqn{g_i = var(X_i) = \lambda\left(\sigma^2 + \left[\kappa\cdot \mu_i\right]^2\right)} and
\item test precision profile model: \eqn{h_i = var(Y_i) = \sigma^2 + \left[\kappa\cdot (\alpha + \beta\mu_i)\right]^2}.
}

The search algorithm implements an efficient approach via reparameterization
to the ratio \eqn{\rho = \frac{\sigma}{\kappa}}.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
n <- 100

quad  <- FALSE   # for a linear fit
quad  <- TRUE    # for a quadratic fit
sigma <- 1
kappa <- 0.08
alpha <- 1
beta  <- 1.1
gamma <- 0
if (quad) gamma <- -0.005
true  <- 8*10^((0:(n-1))/(n-1))
truey <- alpha+beta*true+gamma*true^2
# simulate single sample - set seed for reproducibility
set.seed(1039)
# specifications for predicate method
X     <- sigma*rnorm(100)+true *(1+kappa*rnorm(n))
# specifications for test method
Y     <- sigma*rnorm(100)+truey*(1+kappa*rnorm(n))

# fit with RL precision profile to estimate parameters
RL_gh_fit  <- PWD_get_gh(X,Y, quad=quad)

# results
summary(RL_gh_fit)

}
\references{
Hawkins DM and Kraker JJ (2026). Precision Profile Weighted
Deming Regression for Methods Comparison.
\emph{The Journal of Applied Laboratory Medicine}, \strong{11}(2), 379-392.
\url{doi:10.1093/jalm/jfaf183}

Rocke DM, Lorenzato S (2012). A Two-Component Model for Measurement
Error in Analytical Chemistry.  \emph{Technometrics}, \strong{37:2}:176-184.
}
\author{
Douglas M. Hawkins, Jessica J. Kraker \href{mailto:krakerjj@uwec.edu}{krakerjj@uwec.edu}
}
