% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWD_multi.r
\name{PWD_multi}
\alias{PWD_multi}
\title{Multiple Instrument Weighted Deming Regression}
\usage{
PWD_multi(X, names, refine=1)
}
\arguments{
\item{X}{the \emph{n} by \emph{I} matrix of values reported by the \emph{I} instruments on the \emph{n} samples.}

\item{names}{\emph{optional} (default of NA) - labels of the \emph{I} instruments.}

\item{refine}{\emph{optional} (default of 1) - the number of resubstitutions used to refine lambda estimates.}
}
\value{
A list containing the following components:

\item{L}{the -2 log likelihood}
\item{rho}{the estimate of \eqn{\tho}.}
\item{kappa}{the estimate of \eqn{\kappa}.}
\item{lambda}{the vector of \emph{I} estimated \eqn{\lambda_i}.}
\item{alpha}{the vector of \emph{I} estimated intercepts.}
\item{beta}{the vector of \emph{I} estimated slopes.}
\item{mu}{the vector of \emph{n} estimated true concentrations.}
\item{fity}{the \emph{n} by \emph{I} matrix of fitted values.}
\item{resi}{the \emph{n} by \emph{I} matrix of residuals.}
\item{scalr}{the \emph{n} by \emph{I} matrix of scaled residuals.}
\item{results}{a printer-ready data frame of the fitted regression}
}
\description{
This code estimates the Rocke-Lorenzato precision profile parameters and the
regression of each instrument’s values on the underlying latent true
concentration.
}
\details{
The function optimizes the likelihood over the parameter \code{rho}, calling \code{PWD_innermult} to provide the estimates of the other parameters.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
n  <- 120
I  <- 4
alpha  <- c(0  , 0  , 2  , -2 )
beta   <- c(0.9, 1.1, 1.2, 0.8)
names   <- c("Inst_1", "Inst_2", "Inst_3", "Inst_4")
lambda <- c(1.8, 0.45, 0.9, 0.9)
rutlam <- sqrt(lambda)
sigma  <- 2
kappa  <- 0.08
true   <- 8*10^((0:(n-1))/(n-1))

# simulated data
set.seed(1039)
X     <- NULL
for (i in 1:I)  {
  truey <- alpha[i] + beta[i]*true
  X <- cbind(X, truey+rutlam[i]*(sigma*rnorm(n) + truey*kappa*rnorm(n)))
}

# application
fit <- PWD_multi(X)
with(fit, cat(sprintf("L \%6.3f rho \%6.4g kappa \%6.4f\n", L, rho, kappa)))
print(fit$results)

}
\references{
.
}
\author{
Douglas M. Hawkins, Jessica J. Kraker \href{mailto:krakerjj@uwec.edu}{krakerjj@uwec.edu}
}
