% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_PWD_get_gh.R
\name{summary.pwdgetgh}
\alias{summary.pwdgetgh}
\title{Summary information of output from \code{PWD_get_gh} or \code{multi_PWD}}
\usage{
\method{summary}{pwdgetgh}(object, digits = 6, ...)
}
\arguments{
\item{object}{an object produced by a call to function  \code{PWD_inference} or \code{multi_PWD}}

\item{digits}{\emph{optional} (default of 6) - number of decimal places to use for rounding displayed results.}

\item{...}{Further arguments passed to or from other methods (required for
compatibility with the generic \code{summary} function).}
}
\value{
\code{object} is returned invisibly
per generic \code{summary} method conventions.
}
\description{
The function, invisible to the user, is called by \code{summary()} to display the
formatted results of a model fitted by \code{PWD_inference} or by \code{multi_PWD}.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
quad  <- FALSE   # for a linear fit
quad  <- TRUE    # for a quadratic fit
sigma <- 1
kappa <- 0.08
alpha <- 1
beta  <- 1.1
gamma <- 0
if (quad) gamma <- -0.005
true  <- 8*10^((0:99)/99)
truey <- alpha+beta*true+gamma*true^2
# simulate single sample - set seed for reproducibility
set.seed(1039)
# specifications for predicate method
X     <- sigma*rnorm(100)+true *(1+kappa*rnorm(100))
# specifications for test method
Y     <- sigma*rnorm(100)+truey*(1+kappa*rnorm(100))

# fit with RL precision profile to estimate parameters
RL_gh_fit  <- PWD_get_gh(X,Y, quad=quad)

# results
summary(RL_gh_fit)

}
