% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_PWD_inference.R
\name{summary.pwdinf}
\alias{summary.pwdinf}
\title{Summary information of output from \code{PWD_inference} or \code{multiPWD_inf}}
\usage{
\method{summary}{pwdinf}(object, conflevel = 0.95, digits = 4, ...)
}
\arguments{
\item{object}{an object produced by a call to function  \code{PWD_inference} or \code{multiPWD_inf}}

\item{conflevel}{\emph{optional} (default of 0.95) - confidence level, between 0 to 1;
value of \code{NA} will remove the confidence bounds.}

\item{digits}{\emph{optional} (default of 4) - number of decimal places to use for rounding displayed results.}

\item{...}{Further arguments passed to or from other methods (required for
compatibility with the generic \code{summary} function).}
}
\value{
\code{object} is returned invisibly
per generic \code{summary} method conventions.
}
\description{
The function, invisible to the user, is called by \code{summary()} to display the
formatted results of a model fitted by \code{PWD_inference} or by \code{multiPWD_inf}.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
sigma <- 1
kappa <- 0.08
alpha <- 1
beta  <- 1.1
true  <- 8*10^((0:99)/99)
truey <- alpha+beta*true
# simulate single sample - set seed for reproducibility
set.seed(1039)
# specifications for predicate method
X     <- sigma*rnorm(100)+true *(1+kappa*rnorm(100))
# specifications for test method
Y     <- sigma*rnorm(100)+truey*(1+kappa*rnorm(100))

# fit with RL precision profile to estimate parameters and variability
RL_inf <- PWD_inference(X,Y,MDL=12)
summary(RL_inf)

}
