% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_PWD_outlier.R
\name{summary.pwdout}
\alias{summary.pwdout}
\title{Summary information of output from \code{PWD_outlier} or \code{multi_PWD_out}}
\usage{
\method{summary}{pwdout}(object, ...)
}
\arguments{
\item{object}{an object produced by a call to function  \code{PWD_outlier} or \code{multi_PWD_out}}

\item{...}{Further arguments passed to or from other methods (required for
compatibility with the generic \code{summary} function).}
}
\value{
\code{object} is returned invisibly
per generic \code{summary} method conventions.
}
\description{
The function, invisible to the user, is called by \code{summary()} to display the
formatted results of a model fitted by \code{PWD_outlier} or by \code{multi_PWD_out}.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
sigma <- 1
kappa <- 0.08
alpha <- 1
beta  <- 1.1
true  <- 8*10^((0:99)/99)
truey <- alpha+beta*true
# simulate single sample - set seed for reproducibility
set.seed(1069)
# specifications for predicate method
X     <- sigma*rnorm(100)+true *(1+kappa*rnorm(100))
# specifications for test method
Y     <- sigma*rnorm(100)+truey*(1+kappa*rnorm(100))
# add some outliers
Y[c(1,2,100)] <- Y[c(1,2,100)] + c(-10,9,-50)

# check for outliers, re-fit, and store output
outliers_assess <- PWD_outlier(X, Y, K=5)
summary(outliers_assess)

}
