% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_PWD_resi.R
\name{summary.pwdresi}
\alias{summary.pwdresi}
\title{Summary information of output from \code{PWD_resi}}
\usage{
\method{summary}{pwdresi}(object, digits = 4, ...)
}
\arguments{
\item{object}{an object produced by a call to function  \code{PWD_resi}}

\item{digits}{\emph{optional} (default of 4) - number of decimal places to use for rounding displayed results.}

\item{...}{Further arguments passed to or from other methods (required for
compatibility with the generic \code{summary} function).}
}
\value{
\code{object} is returned invisibly
per generic \code{summary} method conventions.
}
\description{
The function, invisible to the user, is called by \code{summary()} to display the
formatted results of output from \code{PWD_resi}.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
sigma <- 1
kappa <- 0.08
alpha <- 1
beta  <- 1.1
true  <- 8*10^((0:99)/99)
truey <- alpha+beta*true
# simulate single sample - set seed for reproducibility
set.seed(1039)
# specifications for predicate method
X     <- sigma*rnorm(100)+true *(1+kappa*rnorm(100))
# specifications for test method
Y     <- sigma*rnorm(100)+truey*(1+kappa*rnorm(100))

# fit the model and store output
RL_gh_fit  <- PWD_get_gh(X,Y)
# run the residual analysis from the model output
post  <- PWD_resi(X, RL_gh_fit$resi)
summary(post)

}
