% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_PWD_inf.r
\name{multi_PWD_inf}
\alias{multi_PWD_inf}
\title{Jack-knife inference on the parameters of the Multiple Instrument fit}
\usage{
multi_PWD_inf(X, names=NA, refine=1)
}
\arguments{
\item{X}{the \emph{n} by \emph{I} matrix of values reported by the \emph{I} instruments on the \emph{n} samples.}

\item{names}{\emph{optional} (default of NA) - labels of the \emph{I} instruments.}

\item{refine}{\emph{optional} (default of 1) - the number of re-substitutions used to refine lambda estimates.}
}
\value{
A list containing the following components:

\item{L}{the -2 log likelihood of the full-sample fit.}
\item{rho}{the full-sample estimate of \eqn{\rho}.}
\item{kappa}{the full-sample estimate of \eqn{\kappa}.}
\item{lambda}{the vector of \emph{I} estimated \eqn{\lambda_i}.}
\item{lambdase}{the vector of \emph{I} standard errors of the estimated \eqn{\lambda_i}.}
\item{alpha}{the vector of \emph{I} estimated intercepts.}
\item{alphase}{the vector of \emph{I} standard errors of the estimated intercepts.}
\item{beta}{the vector of \emph{I} estimated slopes.}
\item{betase}{the vector of \emph{I} standard errors of the estimated slopes.}
\item{mu}{the vector of \emph{n} estimated true concentrations.}
\item{fity}{the \emph{n} by \emph{I} matrix of fitted values.}
\item{resi}{the \emph{n} by \emph{I} matrix of residuals.}
\item{whichmissing}{a logical vector identifying locations of row of X with missing values}
\item{scalr}{the \emph{n} by \emph{I} matrix of scaled residuals.}
\item{names}{labels of the \emph{I} instruments.}
}
\description{
This code uses jackknifing of the Multiple-Instrument fit to get
standard errors of the regression parameters and lambda values.
}
\details{
Uses jackknifing to get standard errors for the slopes \code{beta},
intercepts \code{alpha}, and \code{lambda} values of the \emph{I} instruments.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
n  <- 100
I  <- 4
alpha  <- c(0  , 0  , 2  , -2 )
beta   <- c(0.9, 1.1, 1.2, 0.8)
names   <- c("Inst_1", "Inst_2", "Inst_3", "Inst_4")
lambda <- c(1.8, 0.45, 0.9, 0.9)
rutlam <- sqrt(lambda)
sigma  <- 2
kappa  <- 0.08
true   <- 8*10^((0:(n-1))/(n-1))

# simulated data
set.seed(1039)
X     <- NULL
for (i in 1:I)  {
  truey <- alpha[i] + beta[i]*true
  X <- cbind(X, truey+rutlam[i]*(sigma*rnorm(n) + truey*kappa*rnorm(n)))
}

# application
inf <- multi_PWD_inf(X, names=names)
inf$alpha; inf$beta

# summary
summary(inf)

}
\references{
Hawkins DM and Kraker JJ. Multiple Instrument Methods Comparison
by Precision weighted Deming Regression,
on \emph{Arxiv} (2026) \url{doi:10.48550/arXiv.2607.11776}
}
\author{
Douglas M. Hawkins, Jessica J. Kraker \href{mailto:krakerjj@uwec.edu}{krakerjj@uwec.edu}
}
