% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_PWD_out.r
\name{multi_PWD_out}
\alias{multi_PWD_out}
\title{Locate and test for outliers of the fit}
\usage{
multi_PWD_out(X, K=NA, names=NA, refine=1, Pcut=0.01)
}
\arguments{
\item{X}{the \emph{n} by \emph{I} matrix of values reported by the \emph{I} instruments on the \emph{n} samples.}

\item{K}{\emph{optional} (default of n/20) the number of cases to seek in the forward outlier screen.}

\item{names}{\emph{optional} (default of NA) - labels of the \emph{I} instruments.}

\item{refine}{\emph{optional} (default of 1) - the number of re-substitutions used to refine lambda estimates.}

\item{Pcut}{\emph{optional} (default of 0.01) the Bonferroni significance level for the outlier test.}
}
\value{
A list containing the following components:

\item{ndrop}{the number of statistically significant outliers. }
\item{drop}{the identities of the significant outliers.}
\item{initlis}{the \emph{K} vector of cases identified as potential outliers in forward selection. }
\item{Dilis}{the \emph{K} vector of Mahalanobis distances of each initial suspect.}
\item{Tilis}{the \emph{K} vector of transforms of the Dilis to Hotelling \eqn{T^2}.}
\item{backlis}{the vector of cases re-included in the backward step.}
\item{backP}{their outlier \emph{P} values.}
\item{bonfP}{The Bonferroni \emph{P} values of the remaining suspect cases.}
\item{forward}{a print-ready listing of the forward selection steps.}
\item{backward}{a print-ready listing of the backward re-inclusion steps.}
\item{outlis}{a print-ready list of the significant outliers’ scaled residuals and Bonferroni \emph{P} values.}
}
\description{
The routine uses a conventional multivariate outlier test statistic set,
in the framework of Rosner’s multiple outlier search methodology,
for the Multiple-Instrument fit. The vectors tested as multivariate outliers
are the vectors of scaled residuals of each case.
}
\details{
A forward selection step identifies the \emph{K} sequentially
most extreme cases using scaled residuals from a sequence of fits.
This is done by finding the Mahalanobis distance of each case’s vector
of scaled residuals from the mean vector of all cases currently thought
to be "clean". The case with the largest such distance is moved from
the "clean" list to the "suspect" list and the model refitted using the
new "clean" list.  This is repeated until \code{K} cases are on the suspect list.

Backward re-inclusion follows.  The model is fitted to the clean cases, and
the Hotelling \eqn{T^2} statistic between each suspect and the set of clean
cases is computed.  The case with the minimum \eqn{T^2} is found and its
Bonferroni P value computed.  If it is not significant, the case is moved
from the suspect back to the clean list.
The process is repeated until either all suspects have been exonerated or
a set of statistically significant outliers remains.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
n  <- 120
I  <- 4
alpha  <- c(0  , 0  , 2  , -2 )
beta   <- c(0.9, 1.1, 1.2, 0.8)
names   <- c("Inst_1", "Inst_2", "Inst_3", "Inst_4")
lambda <- c(1.8, 0.45, 0.9, 0.9)
rutlam <- sqrt(lambda)
sigma  <- 2
kappa  <- 0.08
true   <- 8*10^((0:(n-1))/(n-1))

# simulated data
set.seed(1039)
X     <- NULL
for (i in 1:I)  {
  truey <- alpha[i] + beta[i]*true
  X <- cbind(X, truey+rutlam[i]*(sigma*rnorm(n) + truey*kappa*rnorm(n)))
}

# add outliers
X[1,4] <- X[1,4] + 15
X[n,2] <- X[n,2] / 1.5

# application
out <- multi_PWD_out(X, K=4, names=names)

# full summary of process
summary(out)

}
\references{
Hawkins DM and Kraker JJ. Multiple Instrument Methods Comparison
by Precision weighted Deming Regression,
on \emph{Arxiv} (2026) \url{doi:10.48550/arXiv.2607.11776}
}
\author{
Douglas M. Hawkins, Jessica J. Kraker \href{mailto:krakerjj@uwec.edu}{krakerjj@uwec.edu}
}
