% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WD_General.r
\name{WD_General}
\alias{WD_General}
\title{Weighted Deming Regression}
\usage{
WD_General(X, Y, g, h, epsilon=1e-10)
}
\arguments{
\item{X}{the vector of predicate readings,}

\item{Y}{the vector of test readings,}

\item{g}{the vector of variances of the X,}

\item{h}{the vector of variances of the Y,}

\item{epsilon}{\emph{optional} convergence tolerance limit.}
}
\value{
A list containing the following components:

\item{alpha }{the fitted intercept}
\item{beta }{the fitted slope}
\item{cor }{the Pearson correlation between X and Y}
\item{fity }{the vector of predicted Y}
\item{mu }{the vector of estimated latent true values}
\item{resi }{the vector of residuals}
\item{like }{the -2 log likelihood L}
\item{innr }{the number of inner refinement loops executed}
}
\description{
This code fits the weighted Deming regression on predicate readings (X)
and test readings (Y).
}
\details{
For input vectors \code{g} and \code{h} containing the variances of
predicate readings \code{X} and test readings \code{Y}, respectively, iteratively fits
weighted Deming regression.
}
\examples{
# library
library(ppwdeming)

# parameter specifications
alpha <- 1
beta  <- 1.1
true  <- 8*10^((0:99)/99)
truey <- alpha+beta*true
# Loosely motivated by Vitamin D data set
g     <- 4e-16+0.07*true^1.27
h     <- 6e-2+7e-5*truey^2.2

# simulate single sample - set seed for reproducibility
set.seed(1039)
# specifications for predicate method
X     <- true +sqrt(g)*rnorm(100)
# specifications for test method
Y     <- truey+sqrt(h)*rnorm(100)

# fit with to estimate linear parameters
wd_fit <- WD_General(X,Y,g,h)
cat("\nWith given g and h, the estimated intercept is",
    signif(wd_fit$alpha,4), "and the estimated slope is",
    signif(wd_fit$beta,4), "\n")
}
\references{
Linnet K. Evaluation of regression procedures for methods comparison studies,
1993. \emph{Clinical Chemistry}, \strong{39}: 424-432.
}
\author{
Douglas M. Hawkins, Jessica J. Kraker \href{mailto:krakerjj@uwec.edu}{krakerjj@uwec.edu}
}
